backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.28.1
python-dotenv==0.20.0
